[BUGFIX] Streamline checkbox rendering in TYPO3 Backend
This change uses Twitter Bootstrap 5 usages across TYPO3 Backend,
and fixes the toggle / icon toggle switch.

Now the bootstrap-5 based toggle (form-switch) is used,
and can be modified easily with a custom SVG later-on, but is
kept as is for now, in order to make sure it is used consistently
everywhere.

Resolves: #93310
Releases: master
Change-Id: If52d06932e1ee087d9c7d76cfb7d0f0c0e92276a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67461
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>