[BUGFIX] Migrate carousel classes within the MultiStepWizard component
While updating bootstrap to 5.0.0-beta1
(https://review.typo3.org/c/Packages/TYPO3.CMS/+/67215)
The 'a) Carousel "item" to "carousel-item" class migration' was done
for the ext:backend 'Wizard' component but not for the 'MultiStepWizard'
component.
Therefore the 'new form' wizard will not work properly.
This patch migrates the usage of 'item' classes to 'carousel-item'
within the 'MultiStepWizard' component.

Resolves: #93168
Releases: master
Change-Id: I4c2a47147b668694c84ccda1dd01b7fe3e3d7808
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67579
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>