[BUGFIX] Respect auth mode in page module again
If a non-admin does not have permissions to access
a specific CType, disallowed via AD mode, the header
icons, e.g. edit or delete, are now not longer displayed
in the page module (fluid based) to prevent error
messages in FormEngine.

Therefore, recordEditAccessInternals is used, which
was also the case in PageLayoutView and furthermore
already includes the language access check.

Resolves: #93411
Releases: master, 10.4
Change-Id: I6239cd1ace2f354386b5e6f11613f45b42342800
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67631
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>