[BUGFIX] Add autocomplete attribute for LoginRefresh form
The Backend LoginRefresh password field now also
contains "autocomplete=current-password" just like
the login-form.

Sidenote: The attribute (also for the username) must be added manually
via jQuery as there is a strange bug not allowing "autocomplete".

In addition, it fixes a typo for the rsa-encryption to be
proper english, and not needing a special handling for
LoginRefresh.js anymore.

Resolves: #93448
Releases: master
Change-Id: Iedf306380a4dbfb5ec31c37f932f8d3b64daf366
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67654
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>