[TASK] Load TypeScript declarations for sortablejs
This patch adds the TypeScript declaration for the sortablejs library.
Since the imported module has a different naming, the PageRenderer needs
both the new module name ("sortablejs") and the old name ("Sortable") as
this would be a breaking change, otherwise.

In the master patch, the old module name has been declared deprecated.

Resolves: #93454
Releases: master, 10.4
Change-Id: Ia8cc7bb4fb57bc7d55aa3601d6ccd1bfe6e0b54d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67658
Tested-by: core-ci <typo3@b13.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>