[TASK] De-duplicate Tree code for filtering
This change optimizes the SVG Tree implementations (again),
this time focussing on using proper constructors for SVG
tree and its derivatives (SelectTree, PageTree, FileStorageTree).

In addition, the search + filter logic is now moved
into the base class, in order to reduce duplicate code,
and to re-add features (next steps) that were not
implemented in TYPO3 v9 when re-writing the page tree,
such as highlighting filtered results.

In addition, unused properties and settings are
removed.

This change marks one of the final changes for
reworking the SVG Tree implementation, afterwards
allowing us to move towards:
* native DOM events in favor of d3-dispatch
* custom elements instead of wrapper methods
* decouple "top." and "document." based settings into the Container classes

Resolves: #93701
Releases: master
Change-Id: I55733b8c2d0a84ca263ac6e77d7d0bac30877e25
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68332
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>