[TASK] Extend active evaluation of TOTP provider
The TOTP provider is now only treated as active in case
the active flag is set and the shared secret exists.

It's actually really unlikely that no shared secret
exists, while the active flag is set, because this can
not be manipulated by a user, but only by an admin with
direct database access. However, since the TOTP provider
will never work without a shared secret, this property
is now also checked, to prevent unnecessary errors.

The corresponding tests are adjusted to check both
properties independently.

Resolves: #93776
Releases: master
Change-Id: I9fd85bd82d945db827ea29626e63752c553a4180
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68517
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>