[BUGFIX] Fix content element preview for menu types
With the introduction of fluid styled content in #79622,
the former CType `menu` with its subtypes was splitted
into dedicated content types. Since the content element
preview was not adjusted, it still checked for the `menu`
CType, which did not longer exist in Core, leading to no
preview for any menu type was displayed in page module.

The issue is fixed by adjusting the code to check for the
correct content types, making the preview work again.

This also replaces the last usages of LLL:EXT:lang.

Resolves: #93659
Releases: master, 10.4
Change-Id: I55d40d57da5edd7083cc5c37674f3fa85039e8e4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68563
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>