[TASK] Update @types/bootstrap
In #93810 bootstrap was updated to 5.0.0-beta3,
which contains the new "offcanvas" component.

In order to use it in TypeScript, we need to update
the corresponding type definitions.

The new type definition doo not define return types for
bootstrap JQuery methods, therefore we simply avoid
usage of these.
The update additionally revealed one bug in the login
screen, where the login-submit button did not longer
show a loading indicator as $.button('loading') is no
longer available in bootstrap 5.
This issue is fixed as a drive-by change.

Also @types/jqueryui needs to be patched as the tooltip and
button definitions conflict with those from @bootstrap/types.

Executed command:

  yarn add --dev @types/bootstrap
  yarn patch-package @types/jqueryui

Resolves: #93812
Releases: master
Change-Id: Ie58f62e0c31b4c77e28d1118ecb7144d4f539903
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68576
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>