[BUGFIX] Reload page module after deleting content
The page module is now being reloaded when
deleting a content element via the context menu.

Side note: Since there are a lot of functionalities
in the module, e.g. translation, it isn't sufficient to
just remove the element from DOM, like it's done in
recordlist.
The whole module needs to be reloaded / reinitialized.

Resolves: #91394
Releases: master, 10.4
Change-Id: Ibd9dc853ddb356574a0f426244decb6455f54b44
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68813
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>