[TASK] Re-style record list search box
The search box (accessible via the list module in docheader, and
also used inside the Record Selector) is now simplified by
having a main select field, and additional settings as dropdown
available, making the search box - especially in the record selector -
much more easier to understand and to use. In addition, the record
selector search now looks similar to the File Selector (FAL relations),
as the File Selector uses the same markup now as well.

The doc header in the list module does not jump anymore when
choosing the search box, as this is also now moved into the
main document area, as all other modules (belog etc) use it as well.

The search box component is also extracted from the
DatabaseRecordList class, making it easier to re-use.

As the page module does not have the search anymore,
and thus, would rely on EXT:recordlist to just show the
button, the button (and thus, the jumpyness) is removed
there.

Resolves: #93892
Releases: master
Change-Id: I64259f54a79d0faaeefb45f9528d0a0d0c7b2c59
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68753
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>