[BUGFIX] Prevent currentSubScript re-assignment for secondary modules
With the history-synchronization introduced in #93988,
window.currentSubScript has been unconditionally overwritten
after loading a new module in the content frame.
This causes an empty currentSubScript string to be stored
for secondary modules (modules not linked in the module menu),
and results in subsequent faults, when currentSubScript
is used to open a different page with the current module context.

Releases: master
Resolves: #94036
Related: #93988
Change-Id: Id957e72da5eb39437f35d19469dcdbd8a36c2b7e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68959
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>