[BUGFIX] Fix MultiStepWizard functionality
The MultiStepWizard has some obvious programming errors, which
surprisingly didn't cause any problems in TYPO3 v10. It was built
primarily with the multi step wizard of the form module in mind. This
fact reflects on some parts of the code.

As the direct commits to the module didn't change anything dramatically
it's reasonable to conclude the bootstrap v5 update revealed the flaws.

Most notably all 4 slides are now visible in the progress bar. Before
only 3 slides were visible (Step 3 was overridden with finish).

This patch fixes the prev and next logic to work properly with any size
of slides in a clear and understandable way.

Other notable changes:
- The missing progress bar titles in the form module multi step wizard
are added.
- The jQuery slideUp and slideDown of the buttons when going back to
the first step are removed as it serves no meaningful purpose.
- The hard coded forceSelection change when switching to first slide
(=true) and last slide (=false) is removed.

Resolves: #94037
Releases: master
Change-Id: I4f70dd30b3caf29a71491a69b18c6753666ccef4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68960
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jonas Eberle <flightvision@googlemail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>