[BUGFIX] Fix csh popover position
This fixes the popover position after clicking on a csh link.

Side note:
The popover loads in two steps: First it shows a "loading ..." text and
when the content is fetched the popover must be updated. Before it was
achieved by hiding and then immediately showing the popover again. This
doesn't quite work with bootstrap v5 anymore. The popover is stuck in
the middle of the process before the positioning applies.

New method: Use the "setContent" method of the bootstrap popover
instance. This method is not documented, but it can be used.
Nevertheless we should keep an eye on that, as it's not public API.

Resolves: #93984
Resolves: #93152
Releases: master
Change-Id: If96028bd70bead243af1937cd2f87afed3b61c5f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68962
Tested-by: core-ci <typo3@b13.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>