[TASK] Slim down TBE_EDITOR
TBE_EDITOR is a precursor of the current FormEngine with a lot of unused
legacy code. Most of its functions were used in an internal scope and
were not meant to be public API, thus everything that has no use at the
current time of writing this commit has been removed.

Resolves: #94065
Releases: master
Change-Id: If9099734e80b3f458065047555860775e7e02c89
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69021
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>