[TASK] Deprecate "relativeToCurrentScript" in FAL API
The additional argument "relativeToCurrentScript" through
out the File Abstraction Layer is a faulty conceptual
implementation from back in the days.

FAL can only build "relativeToCurrentScript" links for
local drivers (and only in some cases). In addition, we've
successfully migrated all logic in Frontend (Site Handling)
and Backend (URL Routing since TYPO3 v11) to be able to
handle absolute links, so this option should be marked as
deprecated and never be used again, as it does not properly
work if we want a true Abstraction Layer logic.

Resolves: #94193
Releases: master
Change-Id: Ib9ba752399bdc5b385804d580157d918fe90327c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69173
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>