[BUGFIX] Re-calculate z-index of modal and its backdrop
Bootstraps modals are not designed to have multiple instances opened at
once, which was workarounded in TYPO3 a while ago. If a modal overlays
another modal AND is smaller, its backdrop does not overlay the "lower"
modal due to hardcoded z-index values.

With this patch the z-index of both, the modal and its backdrop are
re-calculated depending on the amount of open modals. We use "1000" as
an overall base to circumvent a stuttering UI as Bootstrap uses a z-index
of "1040" for backdrops on initial rendering.
However, this will clash again when at least four modals are open, which
is fine and should never happen - if it does, it at least reveals a
bad UX.

Resolves: #94155
Releases: master
Change-Id: Ia15d0cdba903d9ec6a9ebe8518b07daf5e52e59f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69181
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>