[TASK] Change password reset functionality
In case administrators don't want their users to
initiate the reset password process, they have to
set the global configuration option "passwordReset"
to false.

This however also prevents administrators from
using the cli command or the action in the backend
user module.

Therefore, some things changed in the password
reset implementation:

- `passwordReset` now only affects the login screen
- The CLI command is always enabled (requires shell
  access since the command is not schedulable)
- Initiating password reset in the backend can be
  managed separately with a new user TSconfig option

Also the descriptions of the global configuration
options are adjusted. Actually `passwordResetForAdmins`
previously mentioned that setting this to FALSE will
only affect the login screen, which was not the case.
It always affected all places.

In the end, this enables administrators to be more
precise when configuring the password reset functionality
in their installations.

Resolves: #91496
Releases: master
Change-Id: Ic74080adffbaf55f82189dffddb93fdd601034e2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69268
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>