[BUGFIX] Add returnUrl for Open Documents/Recently Used Documents
A click on open/recently used documents the returnUrl was not set
which led to a blank content frame when the close button on the
record was clicked.

This issue is fixed by adding the EditDocumentController's returnUrl to
the OpenDocs payload and re-using it in the toolbar items. However, this
will not work with already existing entries and a consistent behavior
cannot be assured due to potentially (dis)allowed modules for each
backend user.

Resolves: #94214
Releases: main, 12.4
Change-Id: I29680172735a5c28c46268cd14b7e26dd0997cd9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69360
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: J�rg B�sche <typo3@joergboesche.de>
Reviewed-by: Susi Moog <look@susi.dev>
Tested-by: Susi Moog <look@susi.dev>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>