[BUGFIX] Fix #modulemenu collapsed state on Firefox
Due to different handling of scrollbar positioning and layouting in
user agents (most use the operating system's graphics stack to show
scrollbars, Chrome its own) it could happen that the modulebar icons
were shown partially only.

It should not be relied on the width of scrollbars and if they affect
the user agent's box model or not.

In our case an inner element was pixel-perfectly relying on parent
elements' box sizes since #93353.

This dependency has been removed for the collapsed state.

Resolves: #94633
Releases: master
Change-Id: Ic573ade55af461935f0cada80a98bf3aaaf9a9c1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70066
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>