[BUGFIX] Add necessary type property to reduced data structure
When using a SelectTree in flexform, the FormSelectTreeAjaxController
reduces the data structure array, passed to the tcaSelectTreeAjaxFieldData
FormDataGroup, down to the relevant element only.

This previously lead to deprecation messages as one of the
DataProviders, FlexPrepare, calls the TcaMigration for the
passed data structure array. This then failed for the container,
since it was falsely treated as an invalid TCA column, due to the
missing type property.

This is now fixed by adding the missing "type=array" property
to the reduced data structure.

Resolves: #94720
Releases: master, 10.4
Change-Id: Ib1ee2fe81ea54e17cdd38a39e2359e1cc6cfdfbf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70253
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>