[TASK] Cleanup Clipboard component
In our effort to improve the UX in the record- and
filelist, the clipboard component will receive some
improvements in upcoming patches. For example the
clipboard panel should be a standalone web component,
listening on events, so it could be used in any module.

However, to be able to do so, the Clipboard class
needs to be cleaned up first. Therefore, among other,
this patch does the following:

* Add strict types, type hints and resolve "mixed"
  types, e.g some arguments accepted bool|int instead
  of just bool, while they are just treated as bool.
* Improve naming of used variables
* Simplify some code parts
* Add guard clauses where applicable
* Move the DataHandler command array generation into
  the classes (SimpleDataHandler, FileController), they
  belong to
* Protect methods, only intended for the use in Clipboard
* Remove fileMode property since most of the code already
  explicitly checked for the given table anyways

Resolves: #94731
Releases: master
Change-Id: Ia72bd1960e2028ee659e5f6db9dac19cf6c04207
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70240
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>