[FEATURE] Introduce multi record selection in filelist
A new component "MultiRecordSelection" is added. It's
used first for the filelist module as well as the file
selector. It simply adds a consistent way of selecting
multiple elements (records) and to perform  actions on
this selection. Therefore, the checkboxes are always
displayed at the beginning of each record row. The
checkbox actions, such as "check all" are available
in the header row of the same column via a dropdown
menu. The actions to perform on the selection, e.g.
"delete" or "import", appear once a record is selected.

For the filelist module, this means those actions are
decoupled from the clipboard, making it easier for
editors to work with multiple records.

This also introduces the "Edit marked" action to the
filelist module, which allows to edit the metadata of
all selected files at once.

Resolves: #94906
Releases: master
Change-Id: Id9e2915203049235fc4b503d33a4f1a77646c133
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70515
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>