[BUGFIX] Replace bootstrap button toggle with css based toggle
Bootstrap 5 dropped the support for data-bs-toggle="buttons".
See: https://github.com/twbs/bootstrap/pull/30650
The replacement is to use the class "btn-check" in combination with the
state of input fields. This has been done in the Localization module.
Some other occurrences didn't rely on the bootstrap js and it was
enough to simply remove the data attribute.

Also remove a forgotten console.log debug.

Resolves: #95048
Releases: master
Change-Id: Icd31e7cc0cb9eb1581a7bcfe1448ed7dad24a935
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70832
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>