[BUGFIX] Fix pasting elements in page module
Having a numeric pad selected, pasting of a single
element did previously not work, using the "paste"
button in the page layout due to the missing "pad"
parameter in the ajax call.

This is now fixed by adding the "pad" parameter,
which changes the current pad to "normal" before
executing the paste command.

Resolves: #87118
Releases: master, 10.4
Change-Id: If96859f3ee1fad3c7c6b0f026c420f715b75626f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70964
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>