[BUGFIX] Handle function switch in page module
Switching from "Languages" to "Columns" mode
in the page module, while having "All languages"
selected, the grid was previously empty, due to
a missing fallback.

The special "-1" is no valid language id in the
"Columns" mode and has to be replaced in
the module settings. We therefore fall back
to the default language, which is in this case
already visually selected for the user, since it's
the first module menu entry.

Resolves: #95161
Related: #92460
Releases: master
Change-Id: I68d242d1fc85fcc9fb48daf7d28298ee70fc7fcf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70981
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>