[BUGFIX] Avoid endless loading of FormEngine
EditDocumentController failed to properly track if
the edit form could be created or not. This lead
to a couple of consequential errors. For example,
the spinner was added but could never be removed,
since the FormEngine JavaScript component was not
initialized.

With PHP > 8 this also lead to a lot of warnings
since $this->firstEl is not set, in case the edit
form could not be created.

This is now fixed by properly tracking whether the
edit form was created and whether $this->firstEl
was initialized. In case the edit form could not
be created, an info box notifies the user.

Resolves: #95183
Related: #95052
Releases: master
Change-Id: I5a710b22319f4ce34b86c64d2141b59ba3a7a12a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71023
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>