[BUGFIX] Initialize selected crop variant on modal open
In the cropper modal the initial aspect ratio
is not respected.

This has been fixed by triggering the update method
when the cropper modal is initialized.

Also by using CropperJs' built-in responsive feature
we avoid tinkering with pixel sizes when window is
resized and optimize ui to be more responsive.

Resolves: #95204
Releases: master
Change-Id: I5fe135220ffa88780b10ac38fa2d14e08bdd7a2e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71087
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>