[TASK] Better behaviour for “Unused elements” after a drag & drop
The unused elements in the page module have a specific styling
(orange heading). When the items are drag&drop'ed into an active
colPos, they previously kept their unused styling.

This change gives the "Unused elements" a global warning background
and removes the warning style from the content items so that they don't
stay orange after a drag and drop in a real colPos.

Resolves: #95230
Releases: master
Change-Id: Idaa323166ec2c8f62b6f0fe747e7965910dbd4b8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71103
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jonas Eberle <flightvision@googlemail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jonas Eberle <flightvision@googlemail.com>
Reviewed-by: Benni Mack <benni@typo3.org>