[BUGFIX] Clone PageLayoutContext for correct language
In the fluid based page module, when in columns mode
and having a L>0 language selected, the context always
contained L=0 as the current site language. Since this
value is used in the fluid templates for data attributes
and similar, it might lead to unintended outcome.

This is now fixed by cloning the context for the currently
selected language before forwarding it to grid creation,
as it's already done in languages mode and for the special
"defLangBinding" view.

Resolves: #95255
Releases: master
Change-Id: I3f6f6213e90123ab74f2b4e04985909986df9534
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71113
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>