[TASK] Split up "new page" functionality into separate action
Previously, the "pagesOnly" argument was used
to link to the PagePositionMap for new pages.

This is now resolved by using an actual separate
route "db_new_pages" to link to the PagePositionMap
to avoid the "special" handling of "pagesOnly", which
still works but creates a redirect to the new route.

Resolves: #95408
Releases: master
Change-Id: I5f9c22751d21c976fff218d942d8674676a64b2c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71337
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>