[BUGFIX] Relocate menu generation in PageLayoutController
Since #92460, the PageLayoutController has to deal
with the special "ALL languages" option. Therefore,
a couple of checks are performed in menuConfig()
to determine whether the special language should
be applied to the menu or not.

It's therefore necessary to move the actual menu
generation after this determination.

This patch also removes a duplicated call to the
makeLanguageMenu() method.

Additionally, the special language is only
added to the language menu, in case at least
one translation of the current page exists.
Previously, the "available" languages were
checked. This makes no sense, since they
might be available, but not yet used.

Resolves: #95387
Related: #92460
Releases: master
Change-Id: I271c488086a7e1d439611f81d3e05019091dacbd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71342
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>